:truck: move `internal/hash` to `internal/auth/hash`
