+++ internal/auth/hash/hash_test.go
