+++ internal/auth/hash/main.go
